fixed, removed mkcert.sh
