+++ front-vite/Dockerfile
@@ -4,10 +4,6 @@ WORKDIR /app
 
 COPY . .
 
-RUN chmod 744 mkcert.sh
-
-RUN ./mkcert.sh
-
 RUN npm install
 
 CMD ["npm", "run", "dev"]
