--- front-vite/ssl/cert.crt
@@ -1,29 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIE/zCCAuegAwIBAgIULODJm70r6pNW+T2em+EaroZYj3MwDQYJKoZIhvcNAQEL
-BQAwDzENMAsGA1UEAwwEY2VydDAeFw0yNDExMjYyMjE4MTVaFw0yNTExMjYyMjE4
-MTVaMA8xDTALBgNVBAMMBGNlcnQwggIiMA0GCSqGSIb3DQEBAQUAA4ICDwAwggIK
-AoICAQDKn5aaSuTtjyL2goxMBnR2a0pJfJZ4/+9vzyKwuXLatYzvj7FukWRC0Fv6
-7+zoJJLjVH7oMMtmq5MpiWhknH2dvvveewZm00dVLFf07bz75yXj+YKla8wyKHK7
-UA0wsatUWYuOCFujVZbOWfKHUPv5BibYoRIpaUz0AG7N1ymgVTyWXpzBCs/jsmhP
-DbZy542yvE+2CfzMku1hWMZtLHBYUP+EjJ8m43NADpfepRtdxTP8TRdkuUFa5jIZ
-a/zt/a5PYyg61stgZArB+6nIO0xEOijzQj1gzYrwmcWurbYP9LcYanD/hvuwRXkT
-gGu+gWsNrfMR6C3cl2QvTc/8SldHaaVCk0aNDsmPM+fnZ52SLoS6GsxHBnDKOIHu
-88CJDGbQEELCI3Y+xygVLrWTJrbo6MRSf82U0SBTeVz+CcfpLFROGxHtadtI9nYg
-aoqzEKGBYjBU8/ybcuDuwoT1KxAVitV4lwahJ/W3VXvyfGksSuyTqAr1neyy6TiN
-erm45Vs9ELP7yS1KX0L3DZIdumkYrqmJyi03sY6/qeV173xpx1bbYPvwTFIgMdB+
-LX3bhDd635uRcPPeZzrqr7xCOnqhGSTIJMzrKee/K+tXjRhop16VnUyaTC5WhVTw
-R8EAhlxBgz4VVGnRbHcb3KeENKj/ZdKKrwS3fg4YOtVb8hzDtQIDAQABo1MwUTAd
-BgNVHQ4EFgQUDf/s4q3mB3CD2QbybE5aHUZ5NpAwHwYDVR0jBBgwFoAUDf/s4q3m
-B3CD2QbybE5aHUZ5NpAwDwYDVR0TAQH/BAUwAwEB/zANBgkqhkiG9w0BAQsFAAOC
-AgEAWJUJ8ovWsbb4gbMGS2QUKgub7M4ZNgg2rH8V4ZhjhClwbzrjWZ0dWGNIQKEj
-KY2DTdW0qXLQF39yFBdd7jLHwlAJOoTFpqV20Miaz2AT0bbGilOBUkq58pfeFQQU
-VAQYczFKDhULsKGuE2QDn61jUW8ANFjOr57JOuff2zoLc8BhNOwRCzg6xbS7FSHB
-yC5UiO4N8gVBMjAKxNDXVxWKI8B6MJ9kpK1Qe1hL1Pmg5cc/h1v1PdvAVnugIRzF
-2ZGu8JeNnqRtSTsVrFdu3JBH74zJnHzBbILLSPZOIEP1CahEChtejr7I95P1RO8m
-KWHkdXnryvNxyM/fnbOg5S2YgHe7RSkZ+pLltKm92XN4FXBIQCZnJoosG8twWCVG
-anQ9VekZPoy68VU+9TbAslZBQXMBGhrWR02JILS/qbymVLwZQ9gQG81ha50NfkyF
-e27yYigbgGZ3dixmuf/clETjJxP+srrj2j5qMINMEPaKz31MenR6Nl96SPmqR8zP
-62hMwtV8lHV8bv9GPCoCknHNdn1yRddD1NHteGWWhETTaNMsovHsYHv/gg/tnzgM
-KM6ZfS4zoKf4w1LTr3Ey9k/VJbe37g+X6z2REBThZgD3kC9aNTyOj/GQyDD0z56u
-EFo6jGkNvFlZsWEA797bKfQQLysArulY3hln01F+Uu1jg7M=
------END CERTIFICATE-----
