--- front-vite/ssl/cert.key
@@ -1,52 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIJQgIBADANBgkqhkiG9w0BAQEFAASCCSwwggkoAgEAAoICAQDKn5aaSuTtjyL2
-goxMBnR2a0pJfJZ4/+9vzyKwuXLatYzvj7FukWRC0Fv67+zoJJLjVH7oMMtmq5Mp
-iWhknH2dvvveewZm00dVLFf07bz75yXj+YKla8wyKHK7UA0wsatUWYuOCFujVZbO
-WfKHUPv5BibYoRIpaUz0AG7N1ymgVTyWXpzBCs/jsmhPDbZy542yvE+2CfzMku1h
-WMZtLHBYUP+EjJ8m43NADpfepRtdxTP8TRdkuUFa5jIZa/zt/a5PYyg61stgZArB
-+6nIO0xEOijzQj1gzYrwmcWurbYP9LcYanD/hvuwRXkTgGu+gWsNrfMR6C3cl2Qv
-Tc/8SldHaaVCk0aNDsmPM+fnZ52SLoS6GsxHBnDKOIHu88CJDGbQEELCI3Y+xygV
-LrWTJrbo6MRSf82U0SBTeVz+CcfpLFROGxHtadtI9nYgaoqzEKGBYjBU8/ybcuDu
-woT1KxAVitV4lwahJ/W3VXvyfGksSuyTqAr1neyy6TiNerm45Vs9ELP7yS1KX0L3
-DZIdumkYrqmJyi03sY6/qeV173xpx1bbYPvwTFIgMdB+LX3bhDd635uRcPPeZzrq
-r7xCOnqhGSTIJMzrKee/K+tXjRhop16VnUyaTC5WhVTwR8EAhlxBgz4VVGnRbHcb
-3KeENKj/ZdKKrwS3fg4YOtVb8hzDtQIDAQABAoICAANxh7x0zbc/sA4fsLY56f2B
-4Gh8tOrhcYmyEj7SJ3fWQlwA0lJsOHV65EkD9Huk0q0JF1e0GrfBh/kWTjnnw91H
-zNtk6WDF6Y9U1ReR8L72h3Kvn5zwaOJXmjvSXXzjMjkIL+GIr9Qhrpp0sVcqG2c0
-g3V+JwEHoP8RPIyRDsOcSeQjpdG8CkWMMQHz/lbgtBBoyRFX0grahU8ax47nnjMn
-5pJvRTpYD/IeLV9Jxzv7oK3G1CeDwlVHK8GpC4ZDfbajVv5Ry3Fb4vSmZZ2Q455a
-agOLyWgdYdz/GzZwyT/exOZ7Snl1PE6BwViBRAr1R8MJ2lJcFc3Xz9/j7a2Ct2Ms
-dhfPOEchN+zsHrgldp2Ukh3lFrs4ZctfZY3iMBwKZdo7ElipADyVzVUDnjqWBPAD
-LQSeZzfHK4bq05Vahgm5/bQV27ppcDvyEWyykOr2o/9yBvSolnGyDtktpXyi3aFc
-8GIKnr8lr1Sbr8+NDTYI31pPMO73QkwO4YKfMLFTTw+AyFcrXjfJWvojvEfGs9v0
-yOTwwtqDRlqZ4DF28tyvFLM/GtIYPMHoPqc0rKgtjfFfolAGFtwIrzfm2O0qF2jk
-jv6YSIPxQJeIamcKxVafIOVdKHzE+6vwWW3P8nOvAKWx6QzitEJLf1Q2sWlxjXy+
-26qpXB1kbDpWj4IOiXYDAoIBAQDonpRcGOhFWaya6yGvXgToOU2rHziCrJpyJW/X
-a4THuYp9bH58TgBxwF8bNBi1WdeXPVvrwkEb3MDFnHtQmDxoMIeJS+rSZkw0aEFe
-M8YdKblj+FAnyDNjMMJ3/81hzQ1BhynxGx44noxhmDZuESLyTGVwGpv3IhYZ6V6W
-1fcCxFmmsIsQSxeY8zeCE+dzCbd/EcVcEOgEn/O2UD3duWWcZmSRqFSLFtxvgmDw
-hTWSmGzhDRFjXfGN/1MnjwBNbFe66zKWS0y3KTjI0xiuBockK3I2qguOmEgZ0MXs
-E8O4VzxxObvb+LhicAfv0TB4F7IsDmuhoZuBA27a/ch2PItzAoIBAQDe/THFfUp2
-zPsJHUIK0cHp3VsihkF9Ko+vl55gQSqVKhrQ2Skl4dvXY61uEJEp+1nRZM8CKjkl
-1A1sJUwMR3fc+s5uczGyGfZl97Dhia9mIFV7vXFSvFYcBWSurJA/tiplRKSFBf13
-/ekkecW9zDDcfnFRMIrDwtrf4ZMNJQAulM03SHXOvsvkWJVTe0J1BzIGmU8lxyhV
-eYB0PUDotE5WnVYOFVeNbS+IGcnhXIOSJOvAfhNZPbIM4i7lf1G0ePbCZZKwSsFe
-tW0f8IWUXX7v6kMVuykbfmJjJEvA1aVX+f7GST6M07Q2tnluB8hVKIlAkR7et7Kd
-w4/+ba4PQno3AoIBAQCnUgmdBT1tRMdm12JvZc1ZzH4VbGuVctnU6SC9btUfZFI9
-TWDdwK2JJLpEMwDi0TDF1Jjdx+hVRkBS4++/VQF+YAndtAHAPVzrBWKfg9Yo++4h
-/maJdvOfkV8QOatdgcTbMYH1r2/ZghGv0XC1UxYz+sesDzfIV3hWQwqKrHT0RI/H
-2ozst1bN6YQvyGle0cpvWDUBuBdr3DHNHYKijcyM5wnHy7CkK1l7y+g2mVqji4gl
-iOTM+LHbVaBWgAHTru6bwLmuaZNQ3BHAQjcqVf+3AJkywLV3aAnUp9r1qHHLktQj
-DUgXPZxDEhpjYbDfcQvzrJu2syvLbNC8LHeYtIwJAoIBAGWpu+JVD/j/1Ue/fhee
-sDDmSaGA4iSlneKpVnx92WMuUhF7oGX0Bh26BRzUEsq4HKf5tONAytMzTXdLvdMv
-Vjd6ZVk3zaVsbsjoV01MT5pZDYAfy6Wz1fJqhw+X/x/89nE7DWLUa04khgGqPvwc
-Nv6f6xvf7LP5h35gwoFQqDeUgcc9kvBSTnTp6hKodJEUWbtmdA4+MWZewbC9KHWv
-lrOogc5j7om9CNmdIExNuHKNwKn/EOUz2pRDHqxlEqQcLixM0MbtHiIGxxyztMMC
-qVfKTxF+ReyR/1tB9gORpBNrF7vhMtJTxqrXIJ+Ijs48b+7u+ssBpz6EJYcTDDNK
-1JsCggEAA9sKFwSeTy49twQknVOrMmeWDq6uu3USdWIqbGNKnSCEU7SBKAiO8p1+
-IBI/S0BQR/NSpLlTgoUeW0wKBppFIYWV/Da1XluhrSK5mNrtTBPGnCQHXHGI+bEf
-CLe1VsbZnA7gthriyavEKsVmuTgCMHUu3jMxRg9BIfIbWacCC+YTdXYgeWyxzRJv
-TBEBZZUDefM/y9hxaX7LkN1gMgf3y5WlrcQJiWtE6JIp5gwOhO0LshJud4kE+jQu
-T05+NVOetaHLs354CBZLUUYikdD7SDHervjrI0OQeQF6AnzzmaiGQaVJnJSMk2hJ
-mXOK2jx+Z5lHZBuHEiwZ8eEl7MRvPw==
------END PRIVATE KEY-----
