+++ front-vite/vite.config.ts
@@ -1,7 +1,6 @@
 import { defineConfig } from 'vite'
-import fs from 'fs';
 import react from '@vitejs/plugin-react'
-import { checkTLSfiles, makeCertTLS } from './src/createCerts'
+import basicSsl from '@vitejs/plugin-basic-ssl'
 
 
 const envVariables: Record<string, string> = {};
@@ -9,10 +8,12 @@ for (const [key, value] of Object.entries(process.env)) {
     envVariables[`import.meta.env.${key}`] = JSON.stringify(value);
 }
 
-
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react()],
+  plugins: [
+    react(),
+    basicSsl(),
+  ],
   define: envVariables,
   server: {
     watch: {
@@ -21,12 +22,5 @@ export default defineConfig({
     port: parseInt(process.env.PORT_FRONTEND as string, 10),
     strictPort: true,
     host: '0.0.0.0',
-    https: {
-      key: String(fs.readFileSync(process.env.SSL_KEY_PATH as string)),
-      cert: String(fs.readFileSync(process.env.SSL_CERT_PATH as string)),
-      ca: [
-        String(fs.readFileSync(process.env.SSL_CERT_PATH as string)),
-      ],
-    },
   }
 });
